+++ __tests__/setup-jest.js
@@ -6,6 +6,7 @@ global.afterEach( () => {
     DI.bind( 'state:rendering', FsmStates.rendering );
     DI.bind( 'fsm', Fsm );
     DI.bind( 'logger', console );
+    DI.resolve( 'sham-ui:store' ).clear();
     document.querySelector( 'body' ).innerHTML = '';
     jest.restoreAllMocks();
 } );